refactor(database): renames database to prisma-orm
