+++ apps/my-t3-app/next.config.mjs
@@ -7,7 +7,7 @@
 /** @type {import("next").NextConfig} */
 const config = {
   reactStrictMode: true,
-  transpilePackages: ["@retconned/ui","@retconned/database"],
+  transpilePackages: ["@retconned/ui","@retconned/prisma-orm"],
   /**
    * If you have the "experimental: { appDir: true }" setting enabled, then you
    * must comment the below `i18n` config out.
