+++ apps/my-t3-app/src/server/api/trpc.ts
@@ -18,7 +18,7 @@ import { type CreateNextContextOptions } from "@trpc/server/adapters/next";
 import { type Session } from "next-auth";
 
 import { getServerAuthSession } from "@/server/auth";
-import { prisma } from "@retconned/database";
+import { prisma } from "../../../../../packages/prisma-orm";
 
 type CreateContextOptions = {
   session: Session | null;
